+++ data/Pandora.Definitions.ResourceManager/Compute/v2022_03_01/RestorePoints/Model-LinuxPatchSettings.cs
@@ -19,7 +19,7 @@ internal class LinuxPatchSettingsModel
     public LinuxPatchAssessmentModeConstant? AssessmentMode { get; set; }
 
     [JsonPropertyName("automaticByPlatformSettings")]
-    public LinuxVMGuestPatchAutomaticByPlatformSettingsModel? AutomaticByPlatformSettings { get; set; }
+    public LinuxVmGuestPatchAutomaticByPlatformSettingsModel? AutomaticByPlatformSettings { get; set; }
 
     [JsonPropertyName("patchMode")]
     public LinuxVMGuestPatchModeConstant? PatchMode { get; set; }
