+++ data/Pandora.Definitions.ResourceManager/Compute/v2022_03_01/RestorePoints/Model-VmSizeProperties.cs
@@ -10,11 +10,14 @@ using Pandora.Definitions.CustomTypes;
 // Licensed under the MIT License. See NOTICE.txt in the project root for license information.
 
 
-namespace Pandora.Definitions.ResourceManager.Compute.v2022_03_01.VirtualMachineScaleSetVMs;
+namespace Pandora.Definitions.ResourceManager.Compute.v2022_03_01.RestorePoints;
 
 
-internal class LinuxVmGuestPatchAutomaticByPlatformSettingsModel
+internal class VmSizePropertiesModel
 {
-    [JsonPropertyName("rebootSetting")]
-    public LinuxVMGuestPatchAutomaticByPlatformRebootSettingConstant? RebootSetting { get; set; }
+    [JsonPropertyName("vCPUsAvailable")]
+    public int? VCPUsAvailable { get; set; }
+
+    [JsonPropertyName("vCPUsPerCore")]
+    public int? VCPUsPerCore { get; set; }
 }
