+++ data/Pandora.Definitions.ResourceManager/Compute/v2022_03_01/VirtualMachineScaleSets/Model-LinuxPatchSettings.cs
@@ -19,7 +19,7 @@ internal class LinuxPatchSettingsModel
     public LinuxPatchAssessmentModeConstant? AssessmentMode { get; set; }
 
     [JsonPropertyName("automaticByPlatformSettings")]
-    public LinuxVmGuestPatchAutomaticByPlatformSettingsModel? AutomaticByPlatformSettings { get; set; }
+    public LinuxVMGuestPatchAutomaticByPlatformSettingsModel? AutomaticByPlatformSettings { get; set; }
 
     [JsonPropertyName("patchMode")]
     public LinuxVMGuestPatchModeConstant? PatchMode { get; set; }
